% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_data.R
\name{half_circle_MVdata}
\alias{half_circle_MVdata}
\title{Simulate Current Status Data from Half Circle Shaped Failure Time and Uniform Censoring Times}
\usage{
half_circle_MVdata(n, p = 2, tau = 9, test = FALSE)
}
\arguments{
\item{n}{A positive integer (sample size)}

\item{p}{a positive integer indicating the dimension of the covariates. Default value is 2.}

\item{tau}{A non-negative value (defining the support of T and C). Default value is 9.}

\item{test}{Logical indicating whether the function should return in addition also the true (uncensored) failure time, together with Bayes risk, for testing. Default value is False and then only current status data is returned (for training).}
}
\value{
A list containing: (1) the simulated data in the format of a data frame with the columns Z, delta, C, and possibly also T (for testing only), and (2) the Bayes risk which can only be evaluated when test=TRUE (otherwise it is NA).
}
\description{
\code{half_circle_MVdata} returns a list that includes a data frame of size [n,p+2]. The first p columns corresponds to the multivarite covarites Z, the p+1 column is the current status indicator, and the p+2 column is the censoring times. Note that when test=TRUE we get an extra column with the true uncensored failure time T.
}
\examples{
d <- half_circle_MVdata(n=100)
}
