% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_data.R
\name{multi_LN_data}
\alias{multi_LN_data}
\title{Simulate Current Status Data from Log-Normal Failure Time Conditional on Multivariate Covariates}
\usage{
multi_LN_data(n, p = 10, tau = 7, test = FALSE, sigma = 1)
}
\arguments{
\item{n}{A positive integer (sample size)}

\item{p}{a positive integer indictating the dimension of the covariates. Default value is 10.}

\item{tau}{A non-negative value (defining the support of T and C). Default value is 7.}

\item{test}{Logical indicating whether the function should return in addition also the true (uncensored) failure time, together with Bayes risk, for testing. Default value is False and then only current status data is returned (for training).}

\item{sigma}{A numeric value indicating the standard deviation of the log-normal distribution.}
}
\value{
A list containing: (1) the simulated data in the format of a data frame with the columns of  the multivariate covariates Z, delta, C, and possibly also T (for testing only), and (2) the Bayes risk which can only be evaluated when test=TRUE (otherwise it is NA).
}
\description{
\code{multi_LN_data} returns a list that includes a data frame of size [n,p+2]. The first p columns corresponds to the multivarite covarites Z, the p+1 column is the current status indicator, and the p+2 column is the censoring times. Note that when test=TRUE we get an extra column with the true uncensored failure time T.
}
\examples{
d <- multi_LN_data(n=100)
}
