% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{news_pop_data_lasso}
\alias{news_pop_data_lasso}
\title{Online News Popularity Data Set}
\format{
A data frame with 39644 rows and 7 variables:
\describe{
  \item{Z1}{first covariate}
  \item{Z2}{second covariate}
  \item{Z3}{third covariate}
  \item{Z4}{fourth covariate}
  \item{Z5}{fifth covariate}
  \item{Z6}{sixth covariate}
  \item{T}{number of shares}
}
}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Online+News+Popularity}
}
\usage{
news_pop_data_lasso
}
\description{
Online news popularity data set after dimensionality reduction and standardization of the covariates.
}
\keyword{datasets}
