% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_predict.R
\name{predict.KMforCSD}
\alias{predict.KMforCSD}
\title{Kernel Machine for Current Status Data - Prediction}
\usage{
\method{predict}{KMforCSD}(object, newdata)
}
\arguments{
\item{object}{A KMforCSD fitted model. This is the output of \code{\link{KMforCSD}}.}

\item{newdata}{A data frame consisting of the p-dimensional covariates Z, the current status indicator vector delta, and the censoring times C.}
}
\value{
The function returns a data frame consisting of newdata, and the kernel machine predictions.
}
\description{
\code{predict.KMforCSD} returns the KM decision function, together with the test data (which can be equivalent to the training data).
}
\examples{
d <- exp_data(n=100) #training set
sol <- KMforCSD(data=d[[1]]) #training
d_test <- exp_data(n=50) #test set
new_data <- predict(sol,d_test[[1]]) #prediction
decision_function <- new_data$prediction
}
